Add files via upload
Proceedings of 2nd National Conference on Recent Advances in Science and Technology, 2k16 (RAST-2k16), 17-18 Feb. 2016 ISBN : 978-93-83046-31-7

Everyone in the world is not having god gifted sight. Some achieve excellence in their life because of blessing with god gifted sight. Blind users are unintentionally deprived of a full participation in the society because of many hefty restrictions on them. For our campus purpose we thought of developing something so that we can provide aid to some extent to the blind/visually impaired student or person in order to navigate closed building campus. A person is blind does not implies he is restricted from using certain technology and also social as others people.

Thus with a purpose to provide same level of services and spaces among all persons, this work proposes an innovative indoor navigation smartphone based application for providing helpful aid for blind/visually impaired student or person in closed campus or at any other place, this proposal system relies on users smartphone and RFID’s deployed in the environment.